abstracted code to copyright.js; simpletabs.js --> LoginTabs.js switching to LoginTabs.js solved the problem where underscore bar wouldn't move. So now it moves, but some interactivity that simpletabs had needs to be migrated over to LoginTabs.js
