+++ src/components/Copyright.js
@@ -1,0 +1,18 @@
+import React from 'react';
+import Typography from "@material-ui/core/Typography";
+import Link from "@material-ui/core/Link";
+
+function Copyright() {
+    return (
+        <Typography variant="body2" color="textSecondary" align="center">
+            {"Copyright © "}
+            <Link color="inherit" href="https://www.redi-school.org/">
+                ReDi School
+            </Link>{" "}
+            {new Date().getFullYear()}
+            {"."}
+        </Typography>
+    );
+}
+
+export default Copyright;
